// backend/app/src/gen/pb/api/v2/app/room/play/service/PlayService_rb.service.ts
// @generated by protoc-gen-restobuf-nestjs v1 with parameter "target=ts,import_extension=js"
// @generated from file api/v2/app/room/play/service.proto (package api.v2.app.room.play, syntax proto3)
/* eslint-disable */

import { Injectable } from "@nestjs/common";
import type { AddPlaceRequest, AddPlaceResponse, ListPlaceRequest, ListPlaceResponse, MoveCardRequest, MoveCardResponse, ShuffleRequest, ShuffleResponse } from "../service_pb.js";
import { Request, Response } from "express";

@Injectable() 
export abstract class PlayServiceService {
  abstract handleListPlace(input: ListPlaceRequest, req: Request, res: Response): Promise<ListPlaceResponse>;
  abstract handleAddPlace(input: AddPlaceRequest, req: Request, res: Response): Promise<AddPlaceResponse>;
  abstract handleShuffle(input: ShuffleRequest, req: Request, res: Response): Promise<ShuffleResponse>;
  abstract handleMoveCard(input: MoveCardRequest, req: Request, res: Response): Promise<MoveCardResponse>;
}